Initial support for compilation on windows.
Verified by building on:

Windows 10 with the following dependencies:
Visual Studio
Msys2
choco install python llvm bazel
pip install numpy
rust

All rust unit tests pass.
Example must be manually compiled currently
This can be fixed later on

+++ rust/src/context.rs
@@ -76,7 +76,7 @@ impl InferenceContext {
                 inputs.len() as ffi::size_t,
                 outputs.as_ptr(),
                 outputs.len() as ffi::size_t,
-                acceleration_backend.bits(),
+                acceleration_backend.bits() as i32,
                 inference_context.as_mut_ptr(),
             );
 
@@ -194,9 +194,9 @@ pub enum InferError {
 
 bitflags! {
     pub struct AccelerationBackend: u32 {
-        const NONE = ffi::RuneCoralAccelerationBackend__None;
-        const EDGETPU = ffi::RuneCoralAccelerationBackend__Edgetpu;
-        const GPU = ffi::RuneCoralAccelerationBackend__Gpu;
+        const NONE = ffi::RuneCoralAccelerationBackend__None as u32;
+        const EDGETPU = ffi::RuneCoralAccelerationBackend__Edgetpu as u32;
+        const GPU = ffi::RuneCoralAccelerationBackend__Gpu as u32;
     }
 }
